@@ -12,35 +12,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.3.8",
-    "@angular/common": "^13.3.8",
-    "@angular/compiler": "^13.3.8",
-    "@angular/core": "^13.3.8",
-    "@angular/elements": "^13.3.8",
-    "@angular/forms": "^13.3.8",
-    "@angular/platform-browser": "^13.3.8",
-    "@angular/platform-browser-dynamic": "^13.3.8",
-    "@angular/router": "^13.3.8",
-    "@interneuroncic/interneuron-ngx-core-lib": "0.0.60",
-    "ajv": "^8.8.2",
-    "bootstrap": "^4.5.0",
+    "@angular/animations": "^18.1.3",
+    "@angular/common": "^18.1.3",
+    "@angular/compiler": "^18.1.3",
+    "@angular/core": "^18.1.3",
+    "@angular/elements": "^18.1.3",
+    "@angular/forms": "^18.1.3",
+    "@angular/platform-browser": "^18.1.3",
+    "@angular/platform-browser-dynamic": "^18.1.3",
+    "@angular/router": "^18.1.3",
+    "ajv": "^8.17.1",
+    "bootstrap": "^5.3.3",
     "font-awesome": "^4.7.0",
-    "jquery": "^3.4.1",
-    "jwt-decode": "^3.1.2",
-    "moment": "^2.29.3",
-    "ngx-bootstrap": "^8.0.0",
-    "ngx-build-plus": "^13.0.1",
-    "popper.js": "^1.16.1",
-    "rxjs": "^6.6.7",
-    "tslib": "^2.0.0",
-    "webpack-dev-server": "^4.9.0",
-    "zone.js": "~0.11.4"
+    "jquery": "^3.7.1",
+    "jwt-decode": "^4.0.0",
+    "moment": "^2.30.1",
+    "ngx-bootstrap": "^18.0.2",
+    "ngx-build-plus": "^18.0.0",
+    "rxjs": "^7.8.1",
+    "tslib": "^2.6.3",
+    "webpack-dev-server": "^5.0.4",
+    "zone.js": "~0.14.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.5",
-    "@angular/cli": "^13.3.5",
-    "@angular/compiler-cli": "^13.3.8",
-    "@angular/language-service": "^13.3.8",
+    "@angular-devkit/build-angular": "^18.1.3",
+    "@angular/cli": "^18.1.3",
+    "@angular/compiler-cli": "^18.1.3",
+    "@angular/language-service": "^18.1.3",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.12.39",
@@ -54,6 +52,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.6.4"
+    "typescript": "~5.5.4"
   }
 }
