@@ -1,7 +1,7 @@
 //BEGIN LICENSE BLOCK 
 //Interneuron Terminus
 
-//Copyright(C) 2024  Interneuron Limited
+//Copyright(C) 2025  Interneuron Limited
 
 //This program is free software: you can redistribute it and/or modify
 //it under the terms of the GNU General Public License as published by
@@ -18,13 +18,26 @@
 //You should have received a copy of the GNU General Public License
 //along with this program.If not, see<http://www.gnu.org/licenses/>.
 //END LICENSE BLOCK 
-
+/* Interneuron Observation App
+Copyright(C) 2023  Interneuron Holdings Ltd
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+See the
+GNU General Public License for more details.
+You should have received a copy of the GNU General Public License
+along with this program.If not, see<http://www.gnu.org/licenses/>. */
 
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http';
 import { AuthenticationService } from './authentication.service';
 import { AppService } from './app.service';
-import { Observable, from } from 'rxjs';
+import { Observable, firstValueFrom, from } from 'rxjs';
+import { environment } from 'src/environments/environment';
 
 @Injectable({
   providedIn: 'root'
@@ -36,25 +49,28 @@ export class ApirequestService {
 
   public getRequest(uri: string): Observable<any> {
     //comment out before push to framework
-    //return from(this.authService.getToken().then((token) => { return this.callApiGet(token, uri); }));
-
-    return from(this.appService.apiService.getRequest(uri));
+    if (!environment.production)
+      return from(this.authService.getToken().then((token) => { return this.callApiGet(token, uri); }));
+    else
+      return from(this.appService.apiService.getRequest(uri));
 
   }
 
   public postRequest(uri: string, body: any): Observable<any> {
     //comment out before push to framework
-    //return from(this.authService.getToken().then((token) => { return this.callApiPost(token, uri, body); }));
-
-    return from(this.appService.apiService.postRequest(uri, body));
+    if (!environment.production)
+      return from(this.authService.getToken().then((token) => { return this.callApiPost(token, uri, body); }));
+    else
+      return from(this.appService.apiService.postRequest(uri, body));
 
   }
 
   public deleteRequest(uri: string): Observable<any> {
     //comment out before push to framework
-    //return from(this.authService.getToken().then((token) => { return this.callApiDelete(token, uri) }));
-
-    return from(this.appService.apiService.deleteRequest(uri));
+    if (!environment.production)
+      return from(this.authService.getToken().then((token) => { return this.callApiDelete(token, uri) }));
+    else
+      return from(this.appService.apiService.deleteRequest(uri));
 
   }
 
@@ -65,8 +81,8 @@ export class ApirequestService {
       'Authorization': 'Bearer ' + token
     });
 
-    return this.httpClient.get(uri, { headers: headers })
-      .toPromise()
+    return firstValueFrom(this.httpClient.get(uri, { headers: headers }))
+     
       .catch((result: HttpErrorResponse) => {
         if (result.status === 401) {
 
@@ -82,8 +98,8 @@ export class ApirequestService {
       'Authorization': 'Bearer ' + token
     });
 
-    return this.httpClient.post(uri, body, { headers: headers })
-      .toPromise()
+    return firstValueFrom(this.httpClient.post(uri, body, { headers: headers }))
+     
       .catch((result: HttpErrorResponse) => {
         console.log(result
         );
@@ -100,8 +116,8 @@ export class ApirequestService {
       'Authorization': 'Bearer ' + token
     });
 
-    return this.httpClient.delete(uri, { headers: headers })
-      .toPromise()
+    return firstValueFrom(this.httpClient.delete(uri, { headers: headers }))
+     
       .catch((result: HttpErrorResponse) => {
         if (result.status === 401) {
 
