@@ -1,7 +1,7 @@
 //BEGIN LICENSE BLOCK 
 //Interneuron Terminus
 
-//Copyright(C) 2024  Interneuron Limited
+//Copyright(C) 2025  Interneuron Limited
 
 //This program is free software: you can redistribute it and/or modify
 //it under the terms of the GNU General Public License as published by
@@ -420,7 +420,7 @@ export class InfusionHistoryComponent implements OnInit,OnDestroy {
   }
 
   CloseCIPopup(){
-    this.subjects.closeCIPopup.next();
+    this.subjects.closeCIPopup.next(true);
   }
 
 }
