@@ -1,7 +1,7 @@
 //BEGIN LICENSE BLOCK 
 //Interneuron Terminus
 
-//Copyright(C) 2024  Interneuron Limited
+//Copyright(C) 2025  Interneuron Limited
 
 //This program is free software: you can redistribute it and/or modify
 //it under the terms of the GNU General Public License as published by
@@ -22,12 +22,14 @@
 export class Uris {
   public baseuri: string;
   public carerecordbaseuri: string;
+  public terminologybaseuri: string;
 }
 
 export class Appsettings {
   sessionStartTime: number;
   maxIntakeRoutes: number;
   maxOutputRoutes: number;
+  ivtype: [];
 }
 
 export class ConfigModel {
