@@ -1,7 +1,7 @@
 //BEGIN LICENSE BLOCK 
 //Interneuron Terminus
 
-//Copyright(C) 2024  Interneuron Limited
+//Copyright(C) 2025  Interneuron Limited
 
 //This program is free software: you can redistribute it and/or modify
 //it under the terms of the GNU General Public License as published by
@@ -33,4 +33,5 @@
     public ispaused : boolean;
     public finishdatetime : Date;
     public _createdDate:Date;
+    public __iscancelled: boolean;
   }
